Added globals.h containing voltage and time calibration for the FADC, changed Archive and DataProcessor classes to use units of voltage and ns instead of ADC bin/channel

+++ src/DataProcessor.h
@@ -34,7 +34,7 @@ public:
 	TH1D* integrate(TH1D* data) const;
 	DataSet* integrateAll(DataSet* data) const;
 	int findMinimumBin(TH1D* data) const;
-	int findDriftTime(const TH1D& data, int threshold) const;
+	int findDriftTime(const TH1D& data, double threshold) const;
 	TH1D* calculateDriftTimeSpectrum(DataSet* data) const;
 
 	void calibrate(TString triggerDataFile);
